+++ yt_xarray/accessor/_readers.py
@@ -18,22 +18,44 @@ def _get_xarray_reader(
         # grid: a yt grid object
         _, fname = field_name
 
-        # this global start index accounts for indexing after any
+        # first get the internal yt index ranges
+        si = grid.get_global_startindex()
+        ei = si + grid.ActiveDimensions
+
+        # convert to the xarray indexing in 3 steps
+        # 1. account for dimension direction
+        # 2. account for any prior xarray subselections
+        # 3. account for interpolation requirements
+
+        # step 1 (if axis has been reversed, node ordering is also reversed)
+        for idim in range(sel_info.ndims):
+            if sel_info.reverse_axis[idim]:
+                # note that the si, ei are exchanged!
+                si0 = si.copy()
+                ei0 = ei.copy()
+                si[idim] = sel_info.global_dims[idim] - ei0[idim]
+                ei[idim] = sel_info.global_dims[idim] - si0[idim]
+
+        # step 2: this global start index accounts for indexing after any
         # subselections on the xarray DataArray are made. might
         # be a way to properly set this with yt grid objects.
         gsi = sel_info.starting_indices
         if sel_info.ndims == 2:
             gsi = np.append(gsi, 0)
+        si = si + gsi
+        ei = ei + gsi
 
-        si = grid.get_global_startindex() + gsi
-        ei = si + grid.ActiveDimensions
+        # step 3: if we are interpolating to cell centers, grab some extra nodes
         if interp_required:
-            # if interpolating, si and ei must be node indices so
-            # we offset by an additional element
-            cell_to_node_offset = np.ones((3,), dtype=int)
-            ei = ei + cell_to_node_offset
+            for idim in range(sel_info.ndims):
+                if sel_info.reverse_axis[idim]:
+                    si[idim] = si[idim] - 1
+                else:
+                    ei[idim] = ei[idim] + 1
 
-        # build the index-selector for our yt grid object
+        # now we can select the data (still accounting for any dimension reversal)
+
+        # build the index-selector for each dimension
         c_list = sel_info.selected_coords  # the xarray coord names
         i_select_dict = {}
         for idim in range(sel_info.ndims):
@@ -62,6 +84,11 @@ def _get_xarray_reader(
         if interp_required:
             # interpolate from nodes to cell centers across all remaining dims
             datavals = _xr_to_yt._interpolate_to_cell_centers(datavals)
+
+        # final flips to account for all the index reversing
+        for idim in range(sel_info.ndims):
+            if sel_info.reverse_axis[idim]:
+                datavals = np.flip(datavals, axis=idim)
 
         # return the plain values
         vals = datavals.values.astype(np.float64)